+++ DashAI/back/src/base/__init__.py
@@ -1,0 +1,2 @@
+from src.base.base_model import BaseModel
+from src.base.base_task import BaseTask
