+++ DashAI/back/src/models/__init__.py
@@ -1,0 +1,5 @@
+from DashAI.back.src.models.sklearn.k_neighbors_classifier import KNeighborsClassifier
+from DashAI.back.src.models.sklearn.random_forest_classifier import (
+    RandomForestClassifier,
+)
+from DashAI.back.src.models.sklearn.svc import SVC
