+++ DashAI/back/src/models/classes/getters.py
