+++ DashAI/back/src/models/enums/squema_types.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
+
 class SquemaTypes(Enum):
     model = "model"
     preprocess = "preprocess"
