+++ DashAI/back/src/models/parameters/models_schemas/knn.json
