+++ DashAI/back/src/models/parameters/models_schemas/svm.json
