+++ DashAI/back/src/models/tabular_classification_model.py
@@ -2,4 +2,4 @@ from models.sklearn.sklearn_model import SklearnModel
 
 
 class TabularClassificationModel(SklearnModel):
-    compatible_tasks = ["TabularClassificationTasks"]
+    _compatible_tasks = ["TabularClassificationTask"]
