+++ DashAI/back/src/registry/__init__.py
@@ -1,0 +1,2 @@
+from registry.model_registry import ModelRegistry
+from registry.task_registry import TaskRegistry
