--- dashai
@@ -1,4 +1,0 @@
-#!/usr/bin/env python
-
-#Esto corre el __init__.py ubicado en la carpeta DashAI
-import DashAI
